refactor: Phase 13b — extract 5 more turn modules to astra-turn-core
- agentic_recursion_guard (35 LOC, zero deps)
- sse_blocks (140 LOC, zero deps)
- parallel_tool_exec (434 LOC, tokio + serde_json)
- streaming_tool_exec (325 LOC, depends on parallel_tool_exec)
- tool_result_sanitize (157 LOC, depends on safety_middleware)
- astra-turn-core: 778 → 812 tests (+34)
- Runtime: 4,052 → 4,018 tests (transferred to crates)
- Added tokio dep to astra-turn-core
