+++ rust/crates/astra-turn-core/src/lib.rs
@@ -5,10 +5,12 @@
 //! tested and compiled independently.
 
 pub mod agent_progress_ui;
+pub mod agentic_recursion_guard;
 pub mod agentic_verdict_audit;
 pub mod boost_domain_hints;
 pub mod cache;
-pub mod cache_diagnostics;pub mod chat_history_openai;
+pub mod cache_diagnostics;
+pub mod chat_history_openai;
 pub mod chat_turn_api_error;
 pub mod chat_turn_explain_wire;
 pub mod confidence_contract;
@@ -37,6 +39,7 @@ pub mod routing;
 pub mod safety_middleware;
 pub mod skill_instructions_merge;
 pub mod snapshot;
+pub mod sse_blocks;
 pub mod sse_edge_stderr_lines;
 pub mod state;
 pub mod stop_hooks;
@@ -44,9 +47,12 @@ pub mod tail_persist;
 pub mod task;
 pub mod tool_args_repair;
 pub mod tool_call_shape;
+pub mod tool_result_sanitize;
 pub mod tool_result_semantics;
 pub mod tool_result_storage;
 pub mod tool_selection;
 pub mod unconsumed;
 pub mod view;
 pub mod xml_tool_call_fallback;
+pub mod parallel_tool_exec;
+pub mod streaming_tool_exec;
